serverless/requirements.txt
python-telegram-bot
futsu>=0.0.21
boto3
